Add lots of documentation for various HTTP objects.
Make MultiMap a first class object and allow it to be required
on its own as with `var MultiMap = require('vertx/multi_map');`.
This is meant for internal use only at the moment, but it seems
reasonable to allow for a constructor that doesn't require an
instance of `org.vertx.java.core.MultiMap` and just creates one
under the covers at construction time.